+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance.hs
@@ -1,10 +1,6 @@
-module Test.Cardano.Ledger.Conformance (
-  module Test.Cardano.Ledger.Conformance.ExecSpecRule.Core,
-  module Test.Cardano.Ledger.Conformance.SpecTranslate.Core,
-  module Test.Cardano.Ledger.Conformance.Utils,
-) where
+module Test.Cardano.Ledger.Conformance (module X) where
 
-import Test.Cardano.Ledger.Conformance.ExecSpecRule.Core
-import Test.Cardano.Ledger.Conformance.Orphans ()
-import Test.Cardano.Ledger.Conformance.SpecTranslate.Core
-import Test.Cardano.Ledger.Conformance.Utils
+import Test.Cardano.Ledger.Conformance.ExecSpecRule.Core as X
+import Test.Cardano.Ledger.Conformance.Orphans as X ()
+import Test.Cardano.Ledger.Conformance.SpecTranslate.Core as X
+import Test.Cardano.Ledger.Conformance.Utils as X